fix: change index type
Change index type from 'int' to 'ptrdiff_t'.

+++ src/selectionsort.c
@@ -4,10 +4,10 @@
 int selectionsort(int *array, ptrdiff_t size) {
     int min_index, temp;
 
-    for (int i = 0; i < size - 1; i++) {
+    for (ptrdiff_t i = 0; i < size - 1; i++) {
         min_index = i;
 
-        for (int j = i + 1; j < size; j++) {
+        for (ptrdiff_t j = i + 1; j < size; j++) {
             if (array[j] < array[min_index]) min_index = j;
         }
